The date columns for the 2018 file is corrupted. AddedPartionedCflux_US-Wkg_HH_201712312330_201812312330.csv
The timestamp for the start and end of the measurement is incorrect and does not increment in a sensible way.
Briefly:
The file starts with 1488/1490 identical timestamps (201801000000), then on the next line is changes (201802000000). The value remains the same for a while and then changes again at line 2833/34 – this is odd because 1488*2 = 2876.

There are 17520 (48*365) lines of data – which suggests that the data is all there. We’ve just been ignoring that year and proceeding with other years.

I took the date columns from the file listed below, separated them and incremented the YEAR by 1.

The two files have identical structures and the same number of lines 365*48.

This six should be vaiable but I've also contacted Russ Scott in case the date error impacted other aspects of the data.

Initial plots indicate the the data is sensible and shows expected patterned of RECO throughout the year

AddedPartionedCflux_US-Wkg_HH_201612312330_201712312330.csv