I've created two new R scripts
CreateVariablePulseDurations.R
reads in the concatonated timeseries of data and creates variable length pulse filters.
It filters the data, using longer pulses for more intense rainstorms.
It creates and writes out dataframes for pulse and non-pulse time.

FitPulseAndNonPulseModels.R
Reads in the dataframes for pulse and non-pulse time created in CreateVariablePulseDurations.R
It then fits Roby's Model4 to pulse and non-pulse time.
It generates model output for both models for the full timeseries and calculates over and underestimates for each model.

I also upload the datafiles - you do not need to run CreateVariablePulseDurations.R  if you don't want to